Fix: Stop shutter widget from accidentally closing shutter
The shutter widget should not be so eagerly modifying shutter
state. It was closing all shutters it interacted with on initialization
and also interfering with autoshutter behavior by closing all shutters
after snapping an image. This is best left to control by the
core AutoShutter mechanism. Particularly a problem as for some
scopes the BF light source is defined as shutter device, so the
widget was turning off a halogen lamp which could never warm up in
time for a fast exposure.