fix: Stop shutter widget from accidentally closing shutter (#121)
* Fix: Stop shutter widget from accidentally closing shutter

The shutter widget should not be so eagerly modifying shutter
state. It was closing all shutters it interacted with on initialization
and also interfering with autoshutter behavior by closing all shutters
after snapping an image. This is best left to control by the
core AutoShutter mechanism. Particularly a problem as for some
scopes the BF light source is defined as shutter device, so the
widget was turning off a halogen lamp which could never warm up in
time for a fast exposure.

* Fix: snap button sends signal on both open+close

Previously it was only checking to send a signal to ensure
that the open signal was processed.

* Fix: only disable the shutter button for the default shutter

* fix: typing

* style(pre-commit.ci): auto fixes [...]

* FIX: Fix shutter button (en/dis)abling

Previously on creation the button for the default shutter
device would not be disabled as it should be if autoshuttering
was on.

* FIX: Don't incorrectly set shutter device on channel change

* FIX: Check shutter state when live mode starts/stops

* TEST: Update tests for new shutter behavior

* FIX: also allow opening shutter on seq finish

* style(pre-commit.ci): auto fixes [...]

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>